Set a user agent
To get the extension version we need to propagate the context to a
helper function.

To propagate that context we need to create the forest instance
dynamically, which is nice because that means we can now trivially
change the patchwork instance URL.

For nodeForSeries to have the 'this' context, we need to call it from a
lambda where this is captured.