Add a way to reply to emails with git send-email
When clicking on the Reply button, a mbox is generated under /tmp and
instructions on how to send this mbox are placed at the top of the file.

This is still a little bit manual (two commands need to be run by the
user) but if this gains momentum we could improve it with a virtual file
system provider that catches saves of the mbox and asks the user
interactively whether they want to send it or not.