Add ability to change Patchwork instance
This change adds a the ability to change the URL of the Patchwork
instance to be used through the extension's settings. All the links
in the source code have been updated to generate an URL based on the
configuration.

Moreover a simple alert and "window reload" prompt is now shown to the
user when the URL is changed.

Signed-off-by: Luca Vizzarro <Luca.Vizzarro@arm.com>